% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cart-to-polar.R
\name{cart_to_polar}
\alias{cart_to_polar}
\title{Convert complex Cartesian coordinates to polar coordinates}
\usage{
cart_to_polar(real, imaginary)
}
\arguments{
\item{real}{A numeric vector, the real part of the complex number}

\item{imaginary}{A numeric vector, the imaginary part of the complex number}
}
\value{
A list with two elements, `r` and `theta`, which correspond to the polar
components of the complex number given in Cartesian coordinates.
}
\description{
Convert complex Cartesian coordinates to polar coordinates
}
\examples{

library(complexR)

# Convert the complex number, z = 1 + 6i to polar form.
cart_to_polar(1, 6)

}
