% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circle-transform.R
\name{circle_transform}
\alias{circle_transform}
\title{Visualize transformations on circles in the complex plane}
\usage{
circle_transform(
  r = 1,
  x0 = 0,
  y0 = 0,
  x_new = expression(x^2 - y^2),
  y_new = expression(2 * x * y),
  theta_min = 0,
  theta_max = 2 * pi
)
}
\arguments{
\item{r}{A numeric vector. The radius of the circle, with default value \eqn{1}.}

\item{x0}{A numeric vector. The x-coordinate of the circle's center, with
default value \eqn{0}.}

\item{y0}{A numeric vector. The y-coordinate of the circle's center, with
default value \eqn{0}.}

\item{x_new}{An object of class expression. The real part of the complex
number obtained from the transformation. The default value is set to
\eqn{expression(x^2 + y^2)}, corresponding to the transformation \eqn{z^2}.}

\item{y_new}{An object of class expression. The imaginary part of the complex
number obtained from the transformation. The default value is set to
\eqn{expression(2xy)}, corresponding to the transformation \eqn{z^2}.}

\item{theta_min}{A numeric vector. The lower bound of the \eqn{\theta} corresponding
to the part of the circle that is part of the domain. The default value is
set to \eqn{0}.}

\item{theta_max}{A numeric vector. The upper bound of the \eqn{\theta} corresponding
to the part of the circle that is part of the domain. The default value is set
to \eqn{2 \pi}.}
}
\value{
A visualization with the domain and the image obtained from the
transformation on all points of the domain.
}
\description{
Visualize transformations on circles in the complex plane
}
\examples{

library(complexR)

# Visualize the transformation z^2 on all points on the unit circle centered at
# (1, 0)
circle_transform(x0 = 1)

# Visualize the transformation 1/z on all points of the first two
# quadrants of the circle of radius 2 centered at (1, 1)
circle_transform(r = 2, x0 = 1, y0 = 1, x_new = expression(x / (x^2 + y^2)),
y_new = expression(-y / (x^2 + y^2)), theta_min = 0, theta_max = pi)

}
