% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{deg_rad}
\alias{deg_rad}
\title{Convert degrees to radians}
\usage{
deg_rad(x)
}
\arguments{
\item{x}{The value in degrees to be converted to radians.}
}
\value{
The value in radians.
}
\description{
Convert degrees to radians
}
\examples{
deg_rad(90)
}
