% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disc-transform.R
\name{disc_transform}
\alias{disc_transform}
\title{Visualize transformations on discs in the complex plane}
\usage{
disc_transform(
  r_inner = 0,
  r_outer = 1,
  x0 = 0,
  y0 = 0,
  x_new = expression(x^2 - y^2),
  y_new = expression(2 * x * y),
  theta_min = 0,
  theta_max = 2 * pi
)
}
\arguments{
\item{r_inner}{A numeric vector. The inner radius of the disc, with default
value \eqn{0}.}

\item{r_outer}{A numeric vector. The outer radius of the disc, with default
value \eqn{1}.}

\item{x0}{A numeric vector. The x-coordinate of the circle's center, with
default value \eqn{0}.}

\item{y0}{A numeric vector. The y-coordinate of the circle's center, with
default value \eqn{0}.}

\item{x_new}{An object of class expression. The real part of the complex
number obtained from the transformation. The default value is set to
\eqn{expression(x^2 + y^2)}, corresponding to the transformation \eqn{z^2}.}

\item{y_new}{An object of class expression. The imaginary part of the complex
number obtained from the transformation. The default value is set to
\eqn{expression(2xy)}, corresponding to the transformation \eqn{z^2}.}

\item{theta_min}{A numeric vector. The lower bound of the \eqn{\theta} corresponding
to the part of the circle that is part of the domain. The default value is
set to \eqn{0}.}

\item{theta_max}{A numeric vector. The upper bound of the \eqn{\theta} corresponding
to the part of the circle that is part of the domain. The default value is set
to \eqn{2 \pi}.}
}
\value{
A visualization with the domain and the image obtained from the
transformation on all points of the domain.
}
\description{
Visualize transformations on discs in the complex plane
}
\examples{

library(complexR)

# Visualize the transformation z^2 on all points on the unit circle centered at
# (1, 0)
disc_transform(x0 = 1)

# Visualize the transformation f(x + iy) = (x^3 - y^2) + i(2*x*y) on all points
# of the first 1/6 of the circle of radius 2 centered at (1, 1)
disc_transform(x0 = 1, y0 = 1, x_new = expression(x^3 - y^2),
y_new = expression(2*x*y), theta_min = 0, theta_max = pi/3)

}
