% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/line-transform.R
\name{line_transform}
\alias{line_transform}
\title{Visualize transformations on straight lines in the complex plane}
\usage{
line_transform(
  m = 1,
  c = 0,
  x_new = expression(x^2 - y^2),
  y_new = expression(2 * x * y),
  xintercept = 1,
  annotations = NULL
)
}
\arguments{
\item{m}{A numeric vector. The slope of the line, with default value set to \eqn{1}.
Set to \eqn{\infty} for horizontal line.}

\item{c}{A numeric vector. The \eqn{y} intercept of the line, with default
value set to \eqn{0}.}

\item{x_new}{An expression. The real part of the complex transformation with
default set to \eqn{x^2 - y^2} corresponding to the \eqn{z^2} transformation.}

\item{y_new}{An expression. The imaginary part of the complex transformation
with default set to \eqn{2xy} corresponding to the \eqn{z^2} transformation.}

\item{x_interecept}{A numeric vector. The \eqn{x}-intercept of the line when
the line is parallel to the \eqn{y}-axis. Default value set to \eqn{1}.}
}
\value{
A visualization with the domain and the image obtained from the
transformation on all points of the domain.
}
\description{
Visualize transformations on straight lines in the complex plane
}
\examples{
library(complexR)

# Visualize the transformation z^2 on the line y = 3x + 5
line_transform(m = 3, c = 5)

}
