% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plane-to-sphere.R
\name{plane_to_sphere}
\alias{plane_to_sphere}
\title{Perform a stereographic projection to project a point on the complex plane to the Riemann sphere}
\usage{
plane_to_sphere(a, b)
}
\arguments{
\item{a}{A numeric vector.}

\item{b}{A numeric vector.}
}
\value{
A list with the coordinates of the point obtained from the projection
of the input to the Riemann sphere.
}
\description{
Perform a stereographic projection to project a point on the complex plane to the Riemann sphere
}
\examples{

library(complexR)

# Convert the complex number z = 1 + 3i to a point on the Riemann sphere

plane_to_sphere(1, 3)

}
