% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polar-to-cart.R
\name{polar_to_cart}
\alias{polar_to_cart}
\title{Convert complex polar coordinates to Cartesian coordinates}
\usage{
polar_to_cart(r, theta)
}
\arguments{
\item{real}{A numeric vector, the radius of the complex number in polar form}

\item{imaginary}{A numeric vector, the angle of the complex number in polar form, between
0 and 2pi.}
}
\value{
A list with two elements, \code{x} and \code{y}, which correspond to the Cartesian
components of the complex number given in polar coordinates.
}
\description{
Convert complex polar coordinates to Cartesian coordinates
}
\examples{

library(complexR)

# Convert the complex number, z = 3(cos(pi/2) + i sin(pi/2)) to Cartesian form.
polar_to_cart(3, pi/2)

}
