% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quad-transform.R
\name{quad_transform}
\alias{quad_transform}
\title{Visualize transformations on squares and rectangles in the complex plane}
\usage{
quad_transform(
  h = 1,
  v = 1,
  x0 = 0,
  y0 = 0,
  x_new = expression(x^2 - y^2),
  y_new = expression(2 * x * y)
)
}
\arguments{
\item{h}{A numeric vector. The breadth of the quadrilateral.}

\item{v}{A numeric vector. The length of the quadrilateral.}

\item{x0}{A numeric vector. The center of the quadrilateral along the real
(\eqn{x-})axis.}

\item{y0}{A numeric vector. The center of the quadrilateral along the imaginary
(\eqn{y-})axis.}

\item{x_new}{An object of class expression. The real part of the complex
number obtained from the transformation. The default value is set to
\eqn{expression(x^2 + y^2)}, corresponding to the transformation \eqn{z^2}.}

\item{y_new}{An object of class expression. The imaginary part of the complex
number obtained from the transformation. The default value is set to
\eqn{expression(2xy)}, corresponding to the transformation \eqn{z^2}.}
}
\description{
Visualize transformations on squares and rectangles in the complex plane
}
