% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rad_deg}
\alias{rad_deg}
\title{Convert radians to degrees}
\usage{
rad_deg(x)
}
\arguments{
\item{x}{The value in radians to be converted to degrees.}
}
\value{
The value in degrees
}
\description{
Convert radians to degrees
}
\examples{
rad_deg(pi/2)
}
