% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\name{rect_polar}
\alias{rect_polar}
\title{Convert from rectangular to polar form}
\usage{
rect_polar(z, ang = "rad")
}
\arguments{
\item{z}{A complex number or vector of complex numbers in rectangular form.}

\item{ang}{The units for the output, can either be rad or deg, defaults to rad.}
}
\value{
The polar form of the complex number as a list, note the units are as specified by ang.
}
\description{
Convert from rectangular to polar form
}
\examples{
rect_polar(1+1i, "deg")
}
