% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{run_app}
\alias{run_app}
\title{Run complexr application}
\usage{
run_app()
}
\value{
Called for its side effect: opens the Shiny application. Returns
invisibly when the app is stopped.
}
\description{
Launches the Shiny app bundled within the package in the default browser.
}
\examples{
\dontrun{
run_app()
}

}
