% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st-plane-plot.R
\name{st_plane_plot}
\alias{st_plane_plot}
\title{Visualize stereographic projections from \eqn{\mathbb{C}^*} to \eqn{\mathbb{C}}}
\usage{
st_plane_plot(x, y, z)
}
\arguments{
\item{x}{A numeric vector. The x-coordinate of the point on the Riemann sphere.}

\item{y}{A numeric vector. The y-coordinate of the point on the Riemann sphere.}

\item{z}{A numeric vector. The z-coordinate of the point on the Riemann sphere.}
}
\value{
A visualization of the point in \eqn{\mathbb{C}^*} and its corresponding
projection in \eqn{\mathbb{C}}.
}
\description{
Visualize stereographic projections from \eqn{\mathbb{C}^*} to \eqn{\mathbb{C}}
}
\examples{

library(complexR)

# Visualize the stereographic projection of the point (1/3, 2/3, 2/3) on the
# complex plane

st_plane_plot(1/3, 2/3, 2/3)

}
