% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st-sphere-plot.R
\name{st_sphere_plot}
\alias{st_sphere_plot}
\title{Visualize stereographic projections from \eqn{\mathbb{C}} to \eqn{\mathbb{C}^*}}
\usage{
st_sphere_plot(a, b)
}
\arguments{
\item{a}{A numeric vector. The real coordinate (\eqn{x}-coordinate) of the number
in \eqn{\mathbb{C}}.}

\item{b}{A numeric vector. The imaginary coordinate (\eqn{y}-coordinate) of the
number in \eqn{\mathbb{C}}.}
}
\value{
A visualization of the point in \eqn{\mathbb{C}} and its corresponding
projection in \eqn{\mathbb{C}^*}.
}
\description{
Visualize stereographic projections from \eqn{\mathbb{C}} to \eqn{\mathbb{C}^*}
}
\examples{

library(complexR)

# Visualize the stereographic projection of x = 2 + 3i on the Riemann sphere
st_sphere_plot(a = 2, b = 3)

}
