% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stereo-proj1.R
\name{stereo_proj1}
\alias{stereo_proj1}
\title{Perform a stereographic projection to project a point on the Riemann sphere to the complex plane}
\usage{
stereo_proj1(x, y, z)
}
\arguments{
\item{x}{A numeric vector.}

\item{y}{A numeric vector.}

\item{z}{A numeric vector.}
}
\value{
The complex number `w` obtained from the projection of the input to the complex plane.
}
\description{
Perform a stereographic projection to project a point on the Riemann sphere to the complex plane
}
\examples{

library(complexR)

# Convert the point (0.5, 0.5, 0.5) on the Riemann sphere to a point on the complex plane

stereo_proj1(0.5, 0.5, 0.5)

}
