% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/i18n.R
\name{i18n_load}
\alias{i18n_load}
\title{Load language dictionary from JSON}
\usage{
i18n_load(lang, path = "app/www/i18n")
}
\arguments{
\item{lang}{character ("es", "en", ...)}

\item{path}{relative path inside inst/}
}
\value{
list
}
\description{
Load language dictionary from JSON
}
