% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/i18n.R
\name{i18n_t}
\alias{i18n_t}
\title{Translate key from dictionary}
\usage{
i18n_t(dict, key)
}
\arguments{
\item{dict}{list}

\item{key}{character (e.g. "mod_datos.title")}
}
\value{
character
}
\description{
Translate key from dictionary
}
