% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results_format.R
\name{plot_results_bar}
\alias{plot_results_bar}
\title{Plot survey estimates as bar chart}
\usage{
plot_results_bar(results)
}
\arguments{
\item{results}{Tibble returned by estimate_survey().}
}
\value{
A ggplot object.
}
\description{
Plot survey estimates as bar chart
}
\examples{
res <- data.frame(
  region   = c("North", "South", "East"),
  estimate = c(0.45, 0.30, 0.55),
  lci      = c(0.40, 0.25, 0.50),
  uci      = c(0.50, 0.35, 0.60),
  estimator = "prop"
)
plot_results_bar(res)

}
