% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\name{polar_rect}
\alias{polar_rect}
\title{Convert from polar to rectangular form}
\usage{
polar_rect(mod, arg)
}
\arguments{
\item{mod}{The modulus of the complex number, or vector of moduli.}

\item{arg}{The argument of the complex number or vector of arguments, in radians.}
}
\value{
The rectangular form of the complex number.
}
\description{
Convert from polar to rectangular form
}
\examples{
polar_rect(1,pi/4)
}
