% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\name{rect_polar}
\alias{rect_polar}
\title{Convert from rectangular to polar form}
\usage{
rect_polar(z)
}
\arguments{
\item{z}{A complex number or vector of complex numbers in rectangular form.}
}
\value{
The polar form of the complex number as a list, note the argument is in radians.
}
\description{
Convert from rectangular to polar form
}
\examples{
rect_polar(1+1i)
}
