% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sphere-to-plane.R
\name{sphere_to_plane}
\alias{sphere_to_plane}
\title{Perform a stereographic projection to project a point on the Riemann sphere to the complex plane}
\usage{
sphere_to_plane(x, y, z)
}
\arguments{
\item{x}{A numeric vector.}

\item{y}{A numeric vector.}

\item{z}{A numeric vector.}
}
\value{
The complex number `w` obtained from the projection of the input to the complex plane.
}
\description{
Perform a stereographic projection to project a point on the Riemann sphere to the complex plane
}
\examples{

library(complexR)

# Convert the point (0.5, 0.5, 0.5) on the Riemann sphere to a point on the complex plane

sphere_to_plane(0.5, 0.5, 0.5)

}
